refactor: collect geometry sanity checks into one
